% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_node_type}
\alias{add_node_type}
\title{Adds gene type to the name of the gene}
\usage{
add_node_type(df)
}
\arguments{
\item{df}{dataframe with all interactions}
}
\description{
Adds ligand, receptor, or transcription factor suffixes to gene names based on their node type.
}
