% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bone_marrow_stromal_cell_example}
\alias{bone_marrow_stromal_cell_example}
\title{Example dataset}
\format{
An object of class \code{Seurat} with 11426 rows and 498 columns.
}
\usage{
bone_marrow_stromal_cell_example
}
\description{
Example dataset for tutorial execution of the package
}
\keyword{datasets}
