% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_r_effectsize}
\alias{calculate_r_effectsize}
\title{Generate cluster and condition heatmap with r effect size only for significant genes}
\usage{
calculate_r_effectsize(
  seuratobject,
  organism,
  out_path,
  celltype_annotation,
  condition_annotation,
  comparison_list
)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{organism}{Organism of sample origin}

\item{out_path}{Output path to save results}

\item{celltype_annotation}{meta data field with celltype annotations}

\item{condition_annotation}{meta data field with condition annotation}

\item{comparison_list}{list of wished comparisons}
}
\value{
tables with r-effectsize results
}
\description{
Description
}
