% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_LR_and_TF_complexes}
\alias{combine_LR_and_TF_complexes}
\title{Combining Ligand-Receptor interaction prediction with Transcription Factor interaction predictions considering receptor complexes}
\usage{
combine_LR_and_TF_complexes(
  tf_table,
  LR_prediction,
  out_path,
  condition,
  add_node_type = FALSE
)
}
\arguments{
\item{tf_table}{table with tf interactions}

\item{LR_prediction}{path to or dataframe with ligand-receptor interaction prediction}

\item{out_path}{path to save results}

\item{condition}{sample condition of data}

\item{add_node_type}{logical whether to add gene type suffixes to gene names}
}
\value{
complete_interactions table with all interactions
}
\description{
Combines ligand-receptor predictions with transcription factor interactions into a CrossTalkeR-compatible network table.
Receptors reported as a complex in the ligand-receptor predictions are matched per subunit and
renamed to the complex, so that the intracellular interactions connect to the ligand-receptor interactions.
}
