% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_LR_and_TF_unfiltered}
\alias{combine_LR_and_TF_unfiltered}
\title{Combining Ligand-Receptor interaction prediction with Transcription Factor interaction predictions}
\usage{
combine_LR_and_TF_unfiltered(tf_table, LR_path, out_path, condition)
}
\arguments{
\item{tf_table}{table with tf interactions}

\item{LR_path}{path to ligand receptor interaction prediction}

\item{out_path}{path to save results}

\item{condition}{sample condition of data}
}
\description{
Description
}
