% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_seurat_to_anndata}
\alias{convert_seurat_to_anndata}
\title{Convert Seurat object to anndata object and save anndata object file}
\usage{
convert_seurat_to_anndata(seuratobject, out_path)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{out_path}{Output path to save results}
}
\description{
Convert Seurat object to anndata object and save anndata object file
}
