% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_empty_CTR_dataframe}
\alias{create_empty_CTR_dataframe}
\title{Create an empty dataframe with CrossTalkeR input table format}
\usage{
create_empty_CTR_dataframe()
}
\description{
Create an empty dataframe with CrossTalkeR input table format
}
