% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{decoupleR_viper_analysis}
\alias{decoupleR_viper_analysis}
\title{Run viper analysis with the decoupleR package}
\usage{
decoupleR_viper_analysis(seuratobject, regulon)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{regulon}{table with weighted TF to target gene interactions}
}
\value{
seuratobject
#' @import decoupleR
}
\description{
Run viper analysis with the decoupleR package
}
