% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dorothea_base_execution}
\alias{dorothea_base_execution}
\title{Run Dorothea based on Seurat object}
\usage{
dorothea_base_execution(
  seuratobject,
  out_path,
  confidence_level = c("A", "B", "C"),
  organism = "human"
)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{out_path}{Output path to save results}

\item{confidence_level}{Curation confidence level to filter DoRothEA regulon (default "ABC")}

\item{organism}{Organism of sample origin}
}
\value{
Seurat object with dorothea assay
}
\description{
Description
}
