% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorothea_tf_prediction.R
\name{dorothea_tf_prediction}
\alias{dorothea_tf_prediction}
\title{Run complete DoRothEA analysis}
\usage{
dorothea_tf_prediction(
  seuratobject,
  out_path,
  confidence_level = c("A", "B", "C"),
  organism = "human",
  condition_Ident,
  celltype_Ident,
  comparison_list,
  pvalue = 0.05,
  log2fc = 0
)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{out_path}{Output path to save results}

\item{confidence_level}{Curation confidence level to filter DoRothEA regulon (default "ABC")}

\item{organism}{Organism of sample origin}

\item{condition_Ident}{Meta data field name that contains condition annotation of cells}

\item{celltype_Ident}{Meta data field name that contains cell type annotation of cells}

\item{comparison_list}{List of condition comparisons}

\item{pvalue}{p-value for filtering}

\item{log2fc}{log2fc value for filtering}
}
\description{
Description
}
