% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_intracellular_networks.R
\name{generate_CrossTalkeR_input}
\alias{generate_CrossTalkeR_input}
\title{Generate CrossTalkeR input from significant tf table}
\usage{
generate_CrossTalkeR_input(
  tf_activities,
  gene_expression,
  regulon = NA,
  organism = "human"
)
}
\arguments{
\item{tf_activities}{Data Frame with transcription factor activities by cell type}

\item{gene_expression}{Table with average gene expression levels}

\item{regulon}{Regulon used to get the transcription factor activities, with the columns source, target and weight}

\item{organism}{Organism for which the data is being processed}
}
\value{
A data frame with CrossTalkeR input
}
\description{
This function loads the transcription factor activity table/data frame
for multiple cell types, and generates the CrossTalkeR input table.
The returned input table contains receptor-transcription factor and
transcription factor-ligand interactions based on the OmniPath database and
the DoRothEA regulon.
}
