% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_CTR_input.R
\name{generate_CrossTalkeR_input_significant_table}
\alias{generate_CrossTalkeR_input_significant_table}
\title{Generate CrossTalkeR input from significant tf table}
\usage{
generate_CrossTalkeR_input_significant_table(
  tf_activities,
  gene_expression,
  regulon = NA
)
}
\arguments{
\item{tf_activities}{Data Frame with transcription factor activities by cell type}

\item{confidence_level}{Confidence Level for the DoRothEA regulon used to get the transcription factor activities}

\item{gene_activities}{Table with average gene expression levels}
}
\value{
A data frame with CrossTalkeR input
}
\description{
This function loads the transcription factor activity table/data frame
for multiple cell types, and generates the CrossTalkeR input table.
The returned input table contains receptor-transcription factor and
transcription factor-ligand interactions based on the OmniPath database and
the DoRothEA regulon.
}
