% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significant_tfs.R
\name{get_significant_tfs}
\alias{get_significant_tfs}
\title{Analyse transcription factor activities for significant transcription factors}
\usage{
get_significant_tfs(
  seuratobject,
  condition,
  out_path,
  pval,
  log2fc,
  tf_condition_significant = NA,
  condition_comparison = FALSE,
  test_type = "t"
)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{condition}{Experminet condition (e.g. disease, knockout ...)}

\item{out_path}{Output path to save results}

\item{pval}{p-value to filter results}

\item{log2fc}{log fold change value to filter results}

\item{tf_condition_significant}{condition comparison results}

\item{condition_comparison}{Whether the condition comparison results should be used to tag the cluster specific results}

\item{test_type}{Statistical test used for the comparison, one of "binom", "t" or "wilcox"}
}
\value{
A data frame with transcription factor activity scores per cell type
}
\description{
Identifies transcription factors with significant activity changes for a given condition and writes summary tables and heatmaps.
}
