% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_significant_tfs.R
\name{get_significant_tfs_single}
\alias{get_significant_tfs_single}
\title{Run DoRothEA analysis to get significant transcription factors}
\usage{
get_significant_tfs_single(seuratobject, condition, out_path, pval, log2fc)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{condition}{Experminet condition (e.g. disease, knockout ...)}

\item{out_path}{Output path to save results}

\item{tf_condition_significant}{condition comparison results}
}
\value{
A data frame with transcription factor activity scores per cell type
}
\description{
Description
}
