% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntraTalker-package.R
\docType{package}
\name{IntraTalker-package}
\alias{IntraTalker}
\alias{IntraTalker-package}
\title{IntraTalker: IntraTalker}
\description{
Workflow for differential transcription factor analysis and generation of intracellular signaling networks.
}
\keyword{internal}
