% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntraTalker.R
\name{IntraTalker_analysis}
\alias{IntraTalker_analysis}
\title{Run IntraTalker analysis}
\usage{
IntraTalker_analysis(seuratobject, tf_activities = NA, arguments_list)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{tf_activities}{matrix with TF activities for each cell in the scRMA-seq data}

\item{arguments_list}{named R list with custom options for the analysis}
}
\value{
TF object containing cluster and condition specific TFs as well as CrossTalkeR input
}
\description{
Perform intracellular network anaysis based on conditions and clusters using decoupleR TF activity results and scRNA-seq data.
}
