% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ligands_homologous}
\alias{ligands_homologous}
\title{Mouse-Human homologous genes table ligands}
\format{
A data frame with 1026 rows and 4 variables:
\itemize{
  \item 10090: Mouse gene symbol
  \item 9606: Human gene symbol
}
}
\usage{
ligands_homologous
}
\description{
A data frame containing the mouse-human homologous genes of the ligands in the Omnipath database
}
\keyword{datasets}
