% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ligands_human}
\alias{ligands_human}
\title{List of annotated ligands from OmnipathR for human data}
\format{
A list with 9708 elements
}
\usage{
ligands_human
}
\description{
List of annotated ligands from OmnipathR for human data
}
\keyword{datasets}
