% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ligands_mouse}
\alias{ligands_mouse}
\title{List of annotated ligands from OmnipathR for mouse data}
\format{
An object of class \code{character} of length 698.
}
\usage{
ligands_mouse
}
\description{
List of annotated ligands from OmnipathR for mouse data
}
\keyword{datasets}
