% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_dorothea_regulon}
\alias{load_dorothea_regulon}
\title{Loading old dorothea package regulons (default if no regulon is provided)}
\usage{
load_dorothea_regulon(organism)
}
\arguments{
\item{organism}{of input data}
}
\value{
regulon
}
\description{
Loading old dorothea package regulons (default if no regulon is provided)
}
