% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouse_regulon_homologous}
\alias{mouse_regulon_homologous}
\title{Mouse-Human homologous genes table transcription factors}
\format{
A data frame with 1026 rows and 4 variables:
\itemize{
  \item 10090: Mouse gene symbol
  \item 9606: Human gene symbol
  \item 10090_ID: Mouse NCBI gene ID
  \item 9606_ID: Human NCBI gene ID
}
}
\usage{
mouse_regulon_homologous
}
\description{
A data frame containing the mouse-human homologous genes of the dorothea regulon
}
\keyword{datasets}
