% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_condition_Heatmaps}
\alias{plot_condition_Heatmaps}
\title{Plotting heatmaps for tf activity comparison between conditions}
\usage{
plot_condition_Heatmaps(
  seuratobject,
  condition_annotation,
  celltype_annotation,
  filter_list,
  out_path
)
}
\arguments{
\item{seuratobject}{seurat object with drothea assay}

\item{condition_annotation}{seurat object meta data field name of conditions}

\item{celltype_annotation}{seurat object meta data field name of cluster}

\item{filter_list}{list of tf genes to filter all detected tf activities}

\item{out_path}{output path for the results}
}
\description{
Description
}
