% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_condition_tf_activities_compact}
\alias{plot_condition_tf_activities_compact}
\title{Plots transcription factor activities compared between cluster and conditions}
\usage{
plot_condition_tf_activities_compact(tf_activity_tables, out_path)
}
\arguments{
\item{tf_activity_tables}{Table with tf activities}

\item{out_path}{Output path to save results}
}
\description{
Creates compact heatmaps of significant transcription factor activity changes for each cluster-condition comparison and combines the resulting PDFs into a single report.
}
