% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_condition_tf_activities_compressed}
\alias{plot_condition_tf_activities_compressed}
\title{Plots transcription factor activities compared between cluster and conditions}
\usage{
plot_condition_tf_activities_compressed(tf_activity_tables, out_path)
}
\arguments{
\item{tf_activity_tables}{Table with tf activities}

\item{out_path}{Output path to save results}
}
\description{
Description
}
