% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_highly_variable_tfs_condition}
\alias{plot_highly_variable_tfs_condition}
\title{Plot heatmap of highly variable transcription factors for conditions}
\usage{
plot_highly_variable_tfs_condition(tf_scores, condition, out_path, clusters)
}
\arguments{
\item{tf_scores}{data frame with transcription factor activity scores per cell type}

\item{condition}{Sample condition for file naming(e.g. control, disease ...)}

\item{out_path}{Output path to save results}

\item{clusters}{Number of clusters}
}
\description{
This function plots a heatmap with the transcription factor activity per
condition. Only the most variable transcription factors are included, which
is determined by the statistical variance of the activity scores.
}
