% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tf_activity_without_mapping_condition}
\alias{plot_tf_activity_without_mapping_condition}
\title{Plots transcription factor activities compared between condition without significance mapping}
\usage{
plot_tf_activity_without_mapping_condition(tf_scores, cluster, out_path)
}
\arguments{
\item{tf_scores}{Table with tf activities}

\item{cluster}{current cluster}

\item{out_path}{Output path to save results}
}
\description{
Description
}
