% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTF_DB}
\alias{RTF_DB}
\title{Dataframe with pre-computed receptor to transcription faactor connections}
\format{
An object of class \code{data.frame} with 29721 rows and 2 columns.
}
\usage{
RTF_DB
}
\description{
Dataframe with pre-computed receptor to transcription faactor connections from Omnipath
}
\keyword{datasets}
