% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTF_DB_2}
\alias{RTF_DB_2}
\title{Dataframe with pre-computed receptor to transcription factor connections 2nd version}
\format{
An object of class \code{data.frame} with 16135 rows and 2 columns.
}
\usage{
RTF_DB_2
}
\description{
Dataframe with pre-computed receptor to transcription factor connections from Omnipath
}
\keyword{datasets}
