% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RTF_DB_mouse}
\alias{RTF_DB_mouse}
\title{Dataframe with pre-computed receptor to transcription factor connections 2nd version for mouse}
\format{
An object of class \code{data.frame} with 14259 rows and 2 columns.
}
\usage{
RTF_DB_mouse
}
\description{
Dataframe with pre-computed receptor to transcription factor connections from Omnipath for mouse
}
\keyword{datasets}
