% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_py_decoupler}
\alias{run_py_decoupler}
\title{Run decoupleR python script.}
\usage{
run_py_decoupler(object_path, regulon_path, out_path)
}
\arguments{
\item{out_path}{Output path to save results}

\item{seuratobject}{Input Seurat Object}
}
\description{
To run decoupleR python version scanpy, pandas, and decoupleR python libraries need to be installed.
}
