% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_py_decoupler_mlm}
\alias{run_py_decoupler_mlm}
\title{Run mlm with decoupleR python script.}
\usage{
run_py_decoupler_mlm(object_path, regulon_path, out_path)
}
\arguments{
\item{object_path}{path to anndata object}

\item{regulon_path}{path to regulon table}

\item{out_path}{Output path to save results}
}
\description{
To run decoupleR python version scanpy, pandas, and decoupleR python libraries need to be installed.
}
