% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_py_liana}
\alias{run_py_liana}
\title{Run mlm with decoupleR python script.}
\usage{
run_py_liana(object_path, regulon_path, out_path)
}
\arguments{
\item{object_path}{path to anndata object}

\item{condition_field}{name of meta-data field with condition information}

\item{cluster_field}{name of meta-data field with cluster information}
}
\description{
To run decoupleR python version scanpy, pandas, and decoupleR python libraries need to be installed.
}
