% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_unfiltered_tf_scores}
\alias{save_unfiltered_tf_scores}
\title{Saves transcription factor activity scores per cell type into a table}
\usage{
save_unfiltered_tf_scores(viper_scores_df, CellsClusters, condition, out_path)
}
\arguments{
\item{viper_scores_df}{data frame with transcription factor activity scores per cell}

\item{CellsClusters}{Clusters perr cell}

\item{condition}{Sample condition for file naming(e.g. control, disease ...)}

\item{out_path}{Output path to save results}
}
\description{
This function saves the transcription factor activity scores per cell type
into a csv table.
}
