% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_variable_tf_scores}
\alias{save_variable_tf_scores}
\title{Saves most variable transcription factor activity scores per cell type into a table}
\usage{
save_variable_tf_scores(tf_scores, condition, out_path)
}
\arguments{
\item{tf_scores}{data frame with transcription factor activity scores per cell type}

\item{condition}{Sample condition for file naming(e.g. control, disease ...)}

\item{out_path}{Output path to save results}
}
\description{
This function saves the transcription factor activity scores per cell type
into a csv table.
}
