% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_activity_analysis.R
\name{tf_activity_analysis}
\alias{tf_activity_analysis}
\title{Run TF activity analysis}
\usage{
tf_activity_analysis(seuratobject, tf_activities = NA, arguments_list)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{tf_activities}{matrix with TF activities for each cell in the scRMA-seq data}

\item{arguments_list}{named R list with custom options for the analysis}
}
\description{
Description
}
