% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfobject.R
\docType{class}
\name{TFObj-class}
\alias{TFObj-class}
\alias{tfobject}
\title{Object structure for the results of the TF activity analyses}
\description{
Object structure for the results of the TF activity analyses
}
\section{Slots}{

\describe{
\item{\code{tf_activities}}{All Cell Cell Interaction Networks}

\item{\code{average_gene_expression}}{All tables from single condition}

\item{\code{regulon}}{Max meanLR from all}

\item{\code{CTR_input}}{All Celltype in the experiment}

\item{\code{intracellular_network}}{Cell Cell Interaction Plots}
}}

