% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_homologous_genes.R
\name{translate_to_human_ligands}
\alias{translate_to_human_ligands}
\title{Translates mouse gene symbols to homologous human gene symbols for ligands}
\usage{
translate_to_human_ligands(ligand_list)
}
\arguments{
\item{ligand_list}{list with ligand genesymbols to be translated}
}
\description{
This function translates mouse gene symbols to homologous human gene symbols
}
