% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_homologous_genes.R
\name{translate_to_human_old}
\alias{translate_to_human_old}
\title{Translates mouse gene symbols to homologous human gene symbols}
\usage{
translate_to_human_old(df)
}
\arguments{
\item{df}{data frame with the transcription factor activities and the gene names as rownames}
}
\description{
This function translates mouse gene symbols to homologous human gene symbols
}
