% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_homologous_genes.R
\name{translate_to_mouse}
\alias{translate_to_mouse}
\title{Translates human gene symbols to homologous mouse gene symbols}
\usage{
translate_to_mouse(df)
}
\arguments{
\item{df}{data frame with the transcription factor activities and the gene names as rownames}
}
\description{
This function translates human gene symbols to homologous mouse gene symbols
}
