% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unfiltered_tf_activity_table}
\alias{unfiltered_tf_activity_table}
\title{Group transcription factor activity by gene and cluster.}
\usage{
unfiltered_tf_activity_table(activity_data_frame, CellsClusters)
}
\arguments{
\item{activity_data_frame}{Data frame with tf activity score by cell}

\item{CellsClusters}{Cluster annotation by cell}
}
\value{
A data frame with transcription factor activity scores per cell type
}
\description{
Description
}
