% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unique_condition_specific_tfs}
\alias{unique_condition_specific_tfs}
\title{Extract all condition specific tfs in one table}
\usage{
unique_condition_specific_tfs(condition_tfs, out_path, log2fc)
}
\arguments{
\item{condition_tfs}{list of tables for condition specific tfs}

\item{out_path}{save path for result table}

\item{log2fc}{log2 fold-change threshold}
}
\description{
Description
}
