% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_entry_to_CTR_dataframe}
\alias{add_entry_to_CTR_dataframe}
\title{Add entry to a dataframe in the CrossTalkeR input format}
\usage{
add_entry_to_CTR_dataframe(
  source,
  target,
  gene_A,
  gene_B,
  type_gene_A,
  type_gene_B,
  MeanLR
)
}
\description{
Add entry to a dataframe in the CrossTalkeR input format
}
