% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_entry_to_Regulon_dataframe}
\alias{add_entry_to_Regulon_dataframe}
\title{Add entry to a dataframe in the CrossTalkeR input format}
\usage{
add_entry_to_Regulon_dataframe(celltype, Receptor, TF, Target_Gene, TF_Score)
}
\description{
Add entry to a dataframe in the CrossTalkeR input format
}
