% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_condition_significant_tfs.R
\name{condition_comparison_significant}
\alias{condition_comparison_significant}
\title{Generate cluster and condition heatmap with r effect size only for significant genes}
\usage{
condition_comparison_significant(
  seuratobject,
  out_path,
  comparison_list,
  num_cell_filter = 0
)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{out_path}{Output path to save results}

\item{comparison_list}{list of wished comparisons}

\item{num_cell_filter}{minimum number of cells in each cell type}
}
\description{
Description
}
