% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_comparison_significant.R
\name{condition_comparison_significant}
\alias{condition_comparison_significant}
\title{Generate cluster and condition heatmap with r effect size only for significant genes}
\usage{
condition_comparison_significant(
  seuratobject,
  out_path,
  celltype_annotation,
  condition_annotation,
  comparison_list
)
}
\arguments{
\item{seuratobject}{Input Seurat Object}

\item{out_path}{Output path to save results}

\item{celltype_annotation}{meta data field with celltype annotations}

\item{condition_annotation}{meta data field with condition annotation}

\item{comparison_list}{list of wished comparisons}
}
\description{
Description
}
