% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_intracellular_networks.R
\name{generate_intracellular_network}
\alias{generate_intracellular_network}
\title{Generate connections in intracellular network}
\usage{
generate_intracellular_network(
  tf_activities,
  gene_expression,
  regulon,
  organism = "human"
)
}
\arguments{
\item{tf_activities}{Data Frame with transcription factor activities by cell type}

\item{confidence_level}{Confidence Level for the DoRothEA regulon used to get the transcription factor activities}

\item{gene_activities}{Table with average gene expression levels}
}
\value{
A data frame with CrossTalkeR input
}
\description{
Loads the transcription factor activity table/data frame
for multiple cell types, and generates a table containing all detected intracellular connections.
}
