% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_tf_activity}
\alias{plot_tf_activity}
\title{Plot heatmap of all transcription factor activities}
\usage{
plot_tf_activity(tf_scores, tag_mapping, condition, out_path)
}
\arguments{
\item{tf_scores}{data frame with transcription factor activity scores per cell type}

\item{tag_mapping}{Labels for heatmap}

\item{condition}{Sample condition for file naming(e.g. control, disease ...)}

\item{out_path}{Output path to save results}
}
\description{
This function plots a heatmap with the transcription factor activities per
cell type.
}
