% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_tf_activity_compact}
\alias{plot_tf_activity_compact}
\title{Plot heatmap of all transcription factor activities compressed version}
\usage{
plot_tf_activity_compact(tf_scores, condition, out_path)
}
\arguments{
\item{tf_scores}{data frame with transcription factor activity scores per cell type}

\item{condition}{Sample condition for file naming(e.g. control, disease ...)}

\item{out_path}{Output path to save results}
}
\description{
Plots a heatmap with the transcription factor activities per
cell type.
}
