% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_input_arguments}
\alias{validate_input_arguments}
\title{Loading old dorothea package regulons (default if no regulon is provided)}
\usage{
validate_input_arguments(arguments_list)
}
\arguments{
\item{arguments_list}{list of user defined arguments}
}
\value{
val_arguments validated list of arguments
}
\description{
Loading old dorothea package regulons (default if no regulon is provided)
}
